Export danger zone as ui-component
